# api/openapi/v1alpha1/query.yaml
openapi: 3.1.0
info:
  title: LLMObs Query API
  version: v1alpha1
  description: >
    The typed read path over the canonical data model, plus the tree/single-entity
    fetches and the score write path (QD-10). The query request body is the DSL
    document (api/query/v1alpha1/dsl.schema.json); entity shapes $ref the canonical
    model JSON Schemas (api/model/v1alpha1/schema/) — one source of truth.
    Ingestion (OTLP) is a different surface and is not defined here.
  license:
    name: Apache-2.0
servers:
  - url: /

# Double-token model (invariant 7): the gateway forwards a kernel-signed service
# token AND a user identity assertion; the Query API computes the permission
# intersection and never trusts a plugin-supplied identity.
security:
  - serviceToken: []
    userAssertion: []
...
paths:
  /v1alpha1/query:
    post:
      tags: [query]
      operationId: runQuery
      summary: Execute a query DSL document.
      description: >
        Row queries for `spans`, `traces`, and `scores` are implemented, as is the
        `traces` score semi-join (QD-9, §8). Traces are derived from spans at query
        time (00-dsl-spec.md §4.1). Aggregation queries (`aggregations`) remain
        `not_implemented` (501) in this maturity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../../query/v1alpha1/dsl.schema.json"
      responses:
        "200":
          description: Query result envelope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryResponse" }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/OverLimit" }
...
  /v1alpha1/spans:
    delete:
      tags: [entities]
      operationId: eraseSpans
      summary: >
        GDPR erasure (requires the `delete` scope). Hard-deletes every span
        matching a mandatory, bounded filter — `user_id` plus a `timeRange` — and
        records an erasure audit row (actor, filter, count, time). Synchronous:
        the filter is bounded and the caller needs the proven count.
      parameters:
        - { name: user_id, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: The erasure result.
          content:
            application/json:
              schema:
                type: object
                required: [erased, audit_id]
                properties:
                  erased: { type: integer }
                  audit_id: { type: string }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "403": { $ref: "#/components/responses/Unauthorized" }
  /v1alpha1/spans/{id}:
    get:
      tags: [entities]
      operationId: getSpan
      summary: Fetch a single span by id.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The span.
          content:
            application/json:
              schema: { $ref: "../../model/v1alpha1/schema/span.schema.json" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1alpha1/scores/{id}:
    get:
      tags: [entities]
      operationId: getScore
      summary: Fetch a single score by id.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The score.
          content:
            application/json:
              schema: { $ref: "../../model/v1alpha1/schema/score.schema.json" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1alpha1/scores:
    post:
      tags: [scores]
      operationId: writeScore
      summary: >
        The score write path (requires the `scores:write` scope). Enforces
        LM-3/LM-8: measurement-only value model and (subject_type, subject_id)
        subjects, including plugin-namespaced subject types. Accepts a single
        score or a batch array; writes are synchronous so validation is
        observable (04-score.md §7). A batch is all-or-nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "../../model/v1alpha1/schema/score.schema.json"
                - type: array
                  items: { $ref: "../../model/v1alpha1/schema/score.schema.json" }
      responses:
        "201":
          description: The number of scores written.
          content:
            application/json:
              schema:
                type: object
                required: [written]
                properties:
                  written: { type: integer }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/OverLimit" }
components:
  securitySchemes:
    serviceToken:
      type: apiKey
      in: header
      name: X-LLMObs-Service-Token
      description: Kernel-signed, short-lived service token identifying the calling plugin.
    userAssertion:
      type: apiKey
      in: header
      name: X-LLMObs-User-Assertion
      description: >
        Kernel-signed identity assertion for the acting user. The Query API grants
        the intersection of the service token's capability and this assertion.
  responses:
    InvalidQuery:
      description: "400: the document fails dsl.schema.json (schema_invalid). Necessary-but-not-sufficient; see 422."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: "403: the effective permission does not grant the requested target/entity (unauthorized)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    OverLimit:
      description: "422: structurally valid but semantically rejected — unknown_field, operator_not_allowed, condition_limit_exceeded, ceiling_exceeded, score_type_mismatch, not_orderable, not_groupable (00-dsl-spec.md §11)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such entity in the caller's project scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    QueryResponse:
      type: object
      additionalProperties: false
      required: [version, data, stats, warnings]
      properties:
        version: { const: v1alpha1 }
        data:
          type: array
          description: >
            Row queries return canonical entities (span/trace/score shapes, subject
            to field-level redaction); aggregation queries return one object per
            group (groupBy values + aggregation aliases).
          items: {}
        cursor:
          type: string
          description: Opaque keyset cursor; present iff a next page exists (row queries only).
        stats:
          type: object
          required: [elapsed_ms]
          properties:
            elapsed_ms: { type: integer, minimum: 0 }
            scanned: { type: integer, minimum: 0 }
            returned: { type: integer, minimum: 0 }
          additionalProperties: true
        warnings:
          type: array
          items: { $ref: "#/components/schemas/Warning" }
    Warning:
      type: object
      additionalProperties: false
      required: [code, message]
      description: >
        Non-fatal signal using the llmobs.dq.* conventions (08-data-quality.md),
        e.g. a page that included truncated payloads or a partial result.
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: object, additionalProperties: true }
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum: [schema_invalid, unknown_field, operator_not_allowed, condition_limit_exceeded, ceiling_exceeded, score_type_mismatch, not_orderable, not_groupable, unauthorized, not_found]
        message: { type: string }
        detail:
          type: object
          additionalProperties: true
          description: Names the offending field/constant where applicable.